AccountSystemLogic/src/main/java/za/ac/nwu/as/logic/flow/impl/CreateAccountTypeFlowImpl.java
package za.ac.nwu.as.logic.flow.impl;


import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;
import za.ac.nwu.as.domain.dto.AccountTypeDto;
import za.ac.nwu.as.logic.flow.CreateAccountTypeFlow;
import za.ac.nwu.as.translator.AccountTypeTranslator;

import javax.transaction.Transactional;
import java.time.LocalDate;

@Transactional
@Component("createAccountTypeFlowName")
public class CreateAccountTypeFlowImpl implements CreateAccountTypeFlow{
    private final AccountTypeTranslator accountTypeTranslator;
    @Autowired
    public CreateAccountTypeFlowImpl(AccountTypeTranslator accountTypeTranslator){
        this.accountTypeTranslator = accountTypeTranslator;
    }

    @Override
    public AccountTypeDto create(AccountTypeDto accountType){
        if(null == accountType.getCreationDate()){
            accountType.setCreationDate(LocalDate.now());
        }
        accountTypeTranslator.someMethod();
       // accountTypeTranslator.getAccountTypeByMnemonic("");
        return accountTypeTranslator.create(accountType);
    }
}
